DB-app/requirements.txt
click>=8.1.3
flask>=2.2.2
flask-sqlalchemy>=3.0.3
itsdangerous>=2.1.2
jinja2>=3.1.2
markupsafe>=2.1.1
mysql-connector-python>=8.3.0
packaging>=21.3
pymysql>=1.0.3
python-dotenv>=1.0.0
sqlalchemy>=2.0.13
tomli>=2.0.1
typing_extensions>=4.5.0
Werkzeug>=2.2.2
flask_marshmallow>=1.2.1
marshmallow-sqlalchemy
cryptography>=36.0.1